feat(dev): :sparkles: create, update and delete todo with user id

+++ src/app/api/todos/route.ts
@@ -1,6 +1,7 @@
 import { NextResponse, NextRequest } from 'next/server';
 import prisma from '../../../lib/prisma';
 import { boolean, object, string } from 'yup';
+import { getServerSessionAction } from '@/auth/actions';
 
 export async function GET(request: Request) {
   const { searchParams } = new URL(request.url);
@@ -33,13 +34,18 @@ const postSchema = object({
 });
 
 export async function POST(request: Request) {
+  const user = await getServerSessionAction();
+
+  if (!user) {
+    return NextResponse.json({ error: 'Unauthorized' }, { status: 401 });
+  }
   try {
     const { complete, description } = await postSchema.validate(
       await request.json()
     );
 
     const todo = await prisma.todo.create({
-      data: { complete, description },
+      data: { complete, description, userId: user.id },
     });
 
     return NextResponse.json(todo);
@@ -49,9 +55,15 @@ export async function POST(request: Request) {
 }
 
 export async function DELETE(request: Request) {
+  const user = await getServerSessionAction();
+
+  if (!user) {
+    return NextResponse.json({ error: 'Unauthorized' }, { status: 401 });
+  }
+
   try {
     const todos = await prisma.todo.deleteMany({
-      where: { complete: true },
+      where: { complete: true, userId: user.id },
     });
 
     return NextResponse.json(todos);
